Deduplicate common utility trait impls.
